Поправлена структура проекта
