+++ README.md
@@ -10,6 +10,6 @@
 2. Добавлен конструктор класса
 ### 3. Создан [utils.py](src/utils.py)
 1. Функция читающая `.json` файлы
-### 4. В [main.py](src/main.py):
+### 4. В [main.py](main.py):
 1. Функция развертывает `list` и передает аргументы в Классы
 2. Переменные передают отдельно в Класс и печатает атрибуты класса
